production_rag/requirements.txt
# Core
chromadb==0.5.5
python-dotenv==1.0.1
tiktoken==0.7.0

# LangChain (LOCKED)
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langgraph==0.2.40
langchain-openai==0.2.0

# Embeddings (local, CPU)
sentence-transformers==2.7.0
rank_bm25==0.2.2

# Document loaders
pypdf==4.2.0

# Optional (only if you really use Cohere)
cohere==5.3.4
langchain-cohere==0.1.5

# Evaluation
ragas==0.4.2
openai==1.30.5

# Backend
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.7.4
pandas==2.2.2
datasets==2.18.0
